feat: step 1 - extract AccessTokenCache to shared token_cache module
