+++ crates/jyc-channels/src/wecom/mod.rs
@@ -14,3 +14,4 @@ pub mod crypto;
 pub mod inbound;
 pub mod outbound;
 pub mod server;
+pub mod token_cache;
